requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
curlify==2.2.1
facebook-business==18.0.5
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pycountry==23.12.11
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.1.0
yarl==1.9.4
